1、增加 service，2 修改 int to long

+++ paicoding-service/src/main/java/com/github/paicoding/forum/service/article/repository/dao/ArticleDao.java
@@ -367,9 +367,9 @@ public class ArticleDao extends ServiceImpl<ArticleMapper, ArticleDO> {
      *
      * @return
      */
-    public Integer countArticle() {
+    public Long countArticle() {
         return lambdaQuery()
                 .eq(ArticleDO::getDeleted, YesOrNoEnum.NO.getCode())
-                .count().intValue();
+                .count();
     }
 }
